requirements.txt
streamlit==1.28.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.41.0
torch==2.1.0
